+++ packages/button/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @poc/button
+# @ChrisVelasco0312/poc-ui-button
+
+## 0.1.0
+
+### Minor Changes
+
+- initial publish
 
 ## 0.1.0