+++ packages/themes/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @poc/themes
+# @ChrisVelasco0312/poc-ui-themes
+
+## 0.1.0
+
+### Minor Changes
+
+- initial publish
 
 ## 0.1.0